Fix requests & proper resource sharing bounds
