+++ scripted/utils.py
@@ -1,17 +1,6 @@
-from pdb import set_trace as T
-
 import nmmo
 from nmmo.lib import material
 
-def l1(start, goal):
-   sr, sc = start
-   gr, gc = goal
-   return abs(gr - sr) + abs(gc - sc)
-
-def l2(start, goal):
-   sr, sc = start
-   gr, gc = goal
-   return 0.5*((gr - sr)**2 + (gc - sc)**2)**0.5
 
 def lInfty(start, goal):
    sr, sc = start
